Renamed shellext.* to utils.*.
