+++ src/shellextension/utils.h
